pyproject.toml
[project]
name = "usigrabber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ijson>=3.4.0.post0",
    "lxml==5.3.2",
    "pandas>=2.3.3",
    "pre-commit>=4.3.0",
    "pyteomics>=4.7.5",
    "requests>=2.32.5",
    "test-package",
    "tqdm>=4.67.1",
    "async-http-client",
    "typer-slim>=0.20.0",
    "python-dotenv>=1.1.1",
    "pydantic>=2.12.3",
    "sqlmodel>=0.0.27",
    "rich>=13.9.4",
    "pronto>=2.7.1",
    "ontology-resolver",
    "psycopg[binary]>=3.2.12",
    "aioftp>=0.27.2",
    "better-exceptions>=0.3.3",
    "cachetools>=6.2.2",
    "enlighten>=1.14.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.44",
]

[project.scripts]
usigrabber = "usigrabber:main"

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = [
    "packages/test-package",
    "packages/async-http-client",
    "packages/ontology-resolver",
]

[tool.uv.sources]
test-package = { workspace = true }
async-http-client = { workspace = true}
ontology-resolver = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "basedpyright>=1.31.7",
    "ruff>=0.14.1",
    "pyinstrument>=5.1.1",
    "py-spy>=0.4.1",
    "memray>=1.19.1",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
    "packages/*/tests",
]
pythonpath = ["."]
addopts = ["--log-cli-level=INFO"]
